% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{as.STFDF}
\alias{as.STFDF}
\alias{as.STFDF,list-method}
\title{ts to STFDF}
\usage{
as.STFDF(timeseries)

\S4method{as.STFDF}{list}(timeseries)
}
\arguments{
\item{timeseries}{Time series.}
}
\description{
This function coerce data from the time series into STFDF data
}
\examples{
chronos = WTSS("http://www.dpi.inpe.br/tws/wtss")
coverages  = listCoverages(chronos)
cv = describeCoverage(chronos, coverages[2])
attributes = cv[[1]]$attributes$name[1] 
ts = timeSeries(chronos, names(cv), attributes, -55,-13,"2000-02-18","2001-01-01")
stfdf <- as.STFDF(ts)
}

