% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{describeCoverage}
\alias{describeCoverage}
\alias{describeCoverage,wtss-method}
\title{Describe coverage}
\usage{
describeCoverage(object, coverages)

\S4method{describeCoverage}{wtss}(object, coverages)
}
\arguments{
\item{object}{A wtss object}

\item{coverages}{A character vector of coverage names}
}
\description{
Describe coverage
}
\examples{
#obj = wtss("http://www.dpi.inpe.br/mds/mds")
#objdesc = describeCoverage(obj,"MOD09Q1")
}

