% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{getListOfTimeSeries}
\alias{getListOfTimeSeries}
\alias{getListOfTimeSeries,wtss-method}
\title{Get list of time series}
\usage{
getListOfTimeSeries(object, coverages, attributes, coordinates, start, end)

\S4method{getListOfTimeSeries}{wtss}(object, coverages, attributes, coordinates,
  start, end)
}
\arguments{
\item{object}{Either a wtss object or a server URL}

\item{coverages}{Either a list of coverages and attributes such as retrieved by describe_coverage() or a character with the coverage name.}

\item{attributes}{A character vector of dataset names.}

\item{coordinates}{A list or data frame of longitude latitude coordinates in WGS84 coordinate system.}

\item{start}{A character with the start date in the format yyyy-mm-dd.}

\item{end}{A character with the end date in the format yyyy-mm-dd.}
}
\description{
This function retrieves the time series for a list of coordinates.
}
\examples{
#obj = wtss("http://www.dpi.inpe.br/ts/wtss")
#objlist = listCoverages(obj)
#objdesc = describeCoverages(obj,objlist)
#coordinates = list( c(longitude=-45, latitude=-12),  c(longitude=-54, latitude=-11))
#tsAll = getListOfTimeSeries(obj, objdesc, coordinates, "2004-01-01", "2004-05-01")
}

