% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{getServerUrl}
\alias{getServerUrl}
\alias{getServerUrl,wtss-method}
\alias{getServerUrl-generic}
\title{Returns the object's server URL}
\usage{
getServerUrl(object)

\S4method{getServerUrl}{wtss}(object)
}
\arguments{
\item{object}{A wtss object}
}
\description{
Returns the object's server URL
}

