% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{list_coverages}
\alias{list_coverages}
\alias{list_coverages,wtss-method}
\title{List coverages}
\usage{
list_coverages(object)

\S4method{list_coverages}{wtss}(object)
}
\arguments{
\item{object}{A wtss object}
}
\description{
List coverages
}
\examples{
#obj = wtssClient("http://www.dpi.inpe.br/mds/mds")
#objlist = listCoverages(obj)
}

