% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{listCoverages}
\alias{listCoverages}
\alias{listCoverages,wtss-method}
\title{List coverages}
\usage{
listCoverages(object)

\S4method{listCoverages}{wtss}(object)
}
\arguments{
\item{object}{A wtss object}
}
\description{
List coverages
}
\examples{
#obj = wtss("http://www.dpi.inpe.br/mds/mds")
#objlist = listCoverages(obj)
}

