% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{listTimeSeries}
\alias{listTimeSeries}
\alias{listTimeSeries,WTSS-method}
\title{Get list of time series}
\usage{
listTimeSeries(object, coverages, attributes, coordinates, start, end)

\S4method{listTimeSeries}{WTSS}(object, coverages, attributes, coordinates,
  start, end)
}
\arguments{
\item{object}{Either a WTSS object or a server URL}

\item{coverages}{Either a list of coverages and attributes such as retrieved by describe_coverage() or a character with the coverage name.}

\item{attributes}{A character vector of dataset names.}

\item{coordinates}{A list or data frame of longitude latitude coordinates in WGS84 coordinate system.}

\item{start}{A character with the start date in the format yyyy-mm-dd or yyyy-mm depending on the coverage.}

\item{end}{A character with the end date in the format yyyy-mm-dd or yyyy-mm depending on the coverage.}
}
\description{
This function retrieves the time series for a list of coordinates.
}
\examples{
#obj = WTSS("http://www.dpi.inpe.br/ts/wtss")
#objlist = listCoverages(obj)
#objdesc = describeCoverage(obj,objlist[2])
#coordinates = list(c(-45,-12),  c(-54,-11))
#tsList = listTimeSeries(obj, names(objdesc), objdesc[[1]]$name, coordinates, "2004-01", "2004-05")
}

