% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{polygonTimeSeries}
\alias{polygonTimeSeries}
\alias{polygonTimeSeries,WTSS-method}
\title{get Time Series by polygon}
\usage{
polygonTimeSeries(object, cv, attributes, polygon, start, end)

\S4method{polygonTimeSeries}{WTSS}(object, cv, attributes, polygon, start, end)
}
\arguments{
\item{object}{Either a WTSS object or a server URL}

\item{cv}{Either a list of coverages and attributes such as retrieved by describe_coverage() or a character with the coverage name.}

\item{attributes}{A character vector of dataset names.}

\item{polygon}{A polygon area.}

\item{start}{A character with the start date in the format yyyy-mm-dd or yyyy-mm depending on the coverage.}

\item{end}{A character with the end date in the format yyyy-mm-dd or yyyy-mm depending on the coverage.}
}
\description{
This function retrieves the time series from a polygon area
}
