% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{setServerUrl}
\alias{setServerUrl}
\alias{setServerUrl,WTSS-method}
\title{Sets the object's server URL}
\usage{
setServerUrl(object, aServerUrl)

\S4method{setServerUrl}{WTSS}(object, aServerUrl)
}
\arguments{
\item{object}{A WTSS object}

\item{aServerUrl}{A character representing the server URL.}
}
\description{
Sets the object's server URL
}
