% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{class}
\name{wtss}
\alias{wtss}
\alias{wtss-class}
\title{The wtss class}
\usage{
wtss(serverUrl)
}
\arguments{
\item{serverUrl}{A server URL}
}
\description{
Use this class for representing a client of a WTSS

Creates a wtss object
}
\note{
No notes
}
\section{Slots }{

\describe{
\item{\code{serverUrl}:}{Object of class \code{"character"}, URL of the server.}
}
}
\examples{
#obj = wtss("http://www.dpi.inpe.br/mds/mds")
}
\author{
Victor Maus, Alber Sanchez, Luiz Fernando Assis, Gilberto Ribeiro
}

