% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{as.STFDF}
\alias{as.STFDF}
\alias{as.STFDF,list-method}
\title{ts to STFDF}
\usage{
as.STFDF(timeseries)

\S4method{as.STFDF}{list}(timeseries)
}
\arguments{
\item{timeseries}{Time series.}
}
\description{
This function coerce data from the time series into STFDF data
}
\examples{
#obj = WTSS("http://www.dpi.inpe.br/tws/wtss")
#objlist = listCoverages(obj)
#objdesc = describeCoverage(obj,objlist[1])
#ts = timeSeries(obj, names(objdesc), "ndvi", -45,-12,"2000-02-18","2004-01-01")
#stfdf <- as.STFDF(ts)
}

