% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{as.STFDF}
\alias{as.STFDF}
\alias{as.STFDF,list-method}
\title{ts to STFDF}
\usage{
as.STFDF(timeseries)

\S4method{as.STFDF}{list}(timeseries)
}
\arguments{
\item{timeseries}{Time series.}
}
\description{
This function coerce data from the time series into STFDF data
}
\examples{
ts.server = WTSS("http://www.dpi.inpe.br/tws/wtss")
coverages  = listCoverages(ts.server)
cv = describeCoverage(ts.server, coverages[5])
attributes = cv[[1]]$attributes$name[1] 
longitude = -47
latitude = -21.7
start_date = "2013-01-01"
end_date =  "2016-12-12"
ts = timeSeries(ts.server, 
                names(cv), 
                attributes, 
                longitude,
                latitude,
                start_date,
                end_date)
stfdf <- as.STFDF(ts)
}
