% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{describe_coverage}
\alias{describe_coverage}
\alias{describe_coverage,wtss-method}
\title{Describe coverage}
\usage{
describe_coverage(object, coverages)

\S4method{describe_coverage}{wtss}(object, coverages)
}
\arguments{
\item{object}{A wtss object}

\item{coverages}{A character vector of coverage names}
}
\description{
Describe coverage
}
\examples{
#obj = wtssClient("http://www.dpi.inpe.br/mds/mds")
#objdesc = describeCoverages(obj,"MOD09Q1")
}

