% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{describeCoverage}
\alias{describeCoverage}
\alias{describeCoverage,WTSS-method}
\title{Describe coverage}
\usage{
describeCoverage(object, coverages)

\S4method{describeCoverage}{WTSS}(object, coverages)
}
\arguments{
\item{object}{A WTSS object}

\item{coverages}{A character vector of coverage names}
}
\description{
Describe coverage
}
\examples{
ts.server = WTSS("http://www.esensing.dpi.inpe.br/wtss/")
cv = describeCoverage(ts.server, c("MOD13Q1"))

}
