% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{getServerUrl}
\alias{getServerUrl}
\alias{getServerUrl-generic}
\alias{getServerUrl,WTSS-method}
\title{Returns the object's server URL}
\usage{
getServerUrl(object)

\S4method{getServerUrl}{WTSS}(object)
}
\arguments{
\item{object}{A WTSS object}
}
\description{
Returns the object's server URL
}
