% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{methods}
\name{listCoverages}
\alias{listCoverages}
\alias{listCoverages,WTSS-method}
\title{List coverages}
\usage{
listCoverages(object)

\S4method{listCoverages}{WTSS}(object)
}
\arguments{
\item{object}{A WTSS object}
}
\description{
List coverages
}
\examples{
ts.server = WTSS("http://www.esensing.dpi.inpe.br/wtss/")
coverages = listCoverages(ts.server)

}
