% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtss.R
\docType{class}
\name{WTSS}
\alias{WTSS}
\alias{WTSS-class}
\alias{WTSS}
\title{The WTSS class}
\usage{
WTSS(serverUrl)
}
\arguments{
\item{serverUrl}{A server URL}
}
\description{
Use this class for creating a connection to a Web Time Series Service (WTSS)

Creates a WTSS object
}
\note{
No notes
}
\section{Slots }{

\describe{
\item{\code{serverUrl}:}{Object of class \code{"character"}, URL of the server.}
\item{\code{listCoverages}:}{Object of class \code{"character"}, list of coverages.}
}
}

\examples{
ts.server = WTSS("http://www.dpi.inpe.br/tws/wtss")
}
\author{
Victor Maus, Alber Sanchez, Luiz Fernando Assis, Pedro Andrade, Gilberto Ribeiro
}
